decode example

+++ src/Types.h
@@ -9,6 +9,6 @@
 #include "ofMain.h"
 
 namespace ofxGraycode {
-	typedef unsigned long ulong;
+	typedef unsigned int uint;
 	typedef unsigned char uchar;
 }
